Fix warnings discovered by msys2' gcc
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38240 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c1541.c
@@ -4584,7 +4584,8 @@ static int write_cmd(int nargs, char **args)
                     if ((endptr - lenptr) > 1) {
                         lenptr[1] = '\0';
                     }
-                    fprintf(stderr, "Converted record length %lu\n", rel_record_length);
+                    fprintf(stderr, "Converted record length %lu\n",
+                            (unsigned long)rel_record_length);
                 } else {
                     rel_record_length = 0;
                 }
